requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.1.0
auto-py-to-exe==2.26.0
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
browsermob-proxy==0.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
Eel==0.14.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
future==0.18.2
geckodriver-autoinstaller==0.1.0
getmac==0.8.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
mysql-connector-python==8.0.31
numpy==1.23.5
openpyxl==3.1.0
outcome==1.2.0
pandas==1.5.2
pefile==2022.5.30
Pillow==9.3.0
protobuf==3.20.1
psutil==5.9.4
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
whichcraft==0.6.1
wsproto==1.2.0
zope.event==4.5.0
zope.interface==5.5.2
